requirements.txt
confluent-kafka
kafka-python
psycopg2-binary
requests
pandas
streamlit
plotly
aiohttp
python-dotenv